Add .github/workflows/release.yml
Official releases would be published to
https://pypi.org/project/ceph-nvmeof/
on every tag releases on github.

Dev releases would be published to
https://test.pypi.org/project/ceph-nvmeof/
on every merge to "devel" branch.

Package will override if the version already exists.
For example, if "1.2.15" already exists on test.pypi
and the new commit also has same version (in pyproject.toml),
then that package would be overwritten (by using a bigger
build number).

Signed-off-by: Vallari Agrawal <val.agl002@gmail.com>